♻️ refactor:product 패키지 생성후 파일 이동, 삭제 + 경로 이름 수정
