+++ src/App.js
@@ -4,8 +4,7 @@ import Auth from "./pages/Auth";
 import Oauth from "./pages/Oauth";
 import "./App.css";
 import ProductPostPage from "./pages/ProductPostPage";
-import ProductDetailPage from "./pages/ProductDetailPage";
-import ProductMain from "./pages/ProductMain";
+import ProductMainPage from "./pages/ProductMainPage";
 
 function App() {
   return (
@@ -13,9 +12,8 @@ function App() {
       <Route path="/" element={<Main />} />
       <Route path="/auth/:id" element={<Auth />} />
       <Route path="/oauth/:id" element={<Oauth />} />
-      <Route path="/products" element={<ProductPostPage />} />
-      <Route path="/products/:id" element={<ProductDetailPage/>} />
-      <Route path="/products" element={<ProductMain />} />
+      <Route path="/productpost" element={<ProductPostPage />} />
+      <Route path="/product" element={<ProductMainPage />} />
     </Routes>
   );
 }
