+++ src/components/main/PostItem.js
@@ -2,10 +2,10 @@ import React from 'react';
 import './popular.css'; // 스타일을 위한 CSS 파일
 import { IoMdHeart } from "react-icons/io";
 
-const PostItem = ({ images, altText, hashtags, views, likeCount }) => {
+const PostItem = ({ imagePath, altText, hashtags, views, likeCount }) => {
   return (
       <div className="item">
-        <img src={images[0].imagePath} alt={altText} />
+        <img src={imagePath} alt={altText} />
         <div className="item-info">
           <div className="hashtags">
             {hashtags.map((hashtag, index) => (
