+++ src/components/product/ProductMain.js
@@ -1,17 +1,9 @@
 import {useEffect, useState} from "react";
-import {setIsLogin} from "../redux/reducer";
 import axios from "axios";
 import {Link, useNavigate} from "react-router-dom";
-import "../pages/productmain.css";
+import "./productmain.css";
 
-// 접속시 보여주는 첫 화면 : /v1/products
-// 드롭박스에서 가격 변경 했을 시 보여주는 화면 : /v1/products?page=${page}&sort=${sortOption}
-// 드롭박스에서 가격 + 지역 변경 했을시 보여주는 화면 : /v1/products?page=${page}&sort=${sortOption}&region=${regionOption}
-// 검색 했을시 기본 이동되는 화면 : ${process.env.REACT_APP_API_URL}/v1/products?page=${page}&search=${searchOption}
-// 검색하고 기본 이동에서 가격 변경했을 시의 화면 : ${process.env.REACT_APP_API_URL}/v1/products?page=${page}&sort=${sortOption}&search=${searchOption}
-// 검색하고 기본 이동 가격변동 + 지역 변경 : ${process.env.REACT_APP_API_URL}/v1/products?page=${page}&sort=${sortOption}&search=${searchOption}&region=${regionOption}
-
-const Product = () => {
+const ProductMain = () => {
   const [page, setPage] = useState(1); // 페이지 상태
   const [sortOption, setSortOption] = useState(undefined); // 정렬 옵션 상태
   const [search, setSearch] = useState(undefined); // 검색어 상태
@@ -21,9 +13,8 @@ const Product = () => {
 
   const navigate = useNavigate();
 
-  const token = localStorage.getItem('accessToken');
 
-  const fetchProducts = async (token, page, sortOption, search,
+  const fetchProducts = async (page, sortOption, search,
       regionOption) => {
     try {
       const response = await axios.get(
@@ -33,9 +24,6 @@ const Product = () => {
               sort: sortOption,
               search: search,
               region: regionOption
-            },
-            headers: {
-              Authorization: token,
             },
             withCredentials: true
           });
@@ -50,12 +38,8 @@ const Product = () => {
 
   useEffect(() => {
 
-    setIsLogin(token);
-
-    if (token) {
-      fetchProducts(token, page, sortOption, search, regionOption);
-    }
-  }, [token, page, sortOption, regionOption]);
+    fetchProducts(page, sortOption, search, regionOption);
+  }, [page, sortOption, regionOption]);
 
   const handlePostClick = (id) => {
     navigate(`/products/${id}`);
@@ -130,7 +114,7 @@ const Product = () => {
   const handleSearchSubmit = (e) => {
     e.preventDefault();
     setPage(1); // 페이지를 1로 설정
-    fetchProducts(token, 1, sortOption, search, regionOption); // 검색어를 포함한 검색 요청
+    fetchProducts(1, sortOption, search, regionOption); // 검색어를 포함한 검색 요청
   };
 
   return (
@@ -194,5 +178,5 @@ const Product = () => {
   );
 }
 
-export default Product;
+export default ProductMain;
