+++ src/components/product/productmain.css
@@ -8,7 +8,7 @@ body {
 
 .container {
   max-width: 1200px;
-  margin: 0 auto;
+  margin: 60px auto;
   padding: 50px 20px 25px;
 }
