--- src/pages/ProductDetailPage.js
@@ -1,10 +1,0 @@
-import ProductDetail from "../components/ProductDetail";
-
-function ProductDetailPage(){
-  return(
-      <div>
-        <ProductDetail />
-      </div>
-  );
-}
-export default ProductDetailPage;
